+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   const {StringElement, ArrayElement, ObjectElement} = namespace.elements;
   ```
 
+- Add functionality for Embedded Refract
+
 # 0.11.0 - 2015-09-07
 
 - **BREAKING** The public interface of the `minim` module has changed significantly. List of changes: